refactor(application): Try change the scrutucture for many libs and connected all.
