--- mottu-mapping-api-dotnet.http
@@ -1,6 +1,0 @@
-@mottu_mapping_api_dotnet_HostAddress = http://localhost:5184
-
-GET {{mottu_mapping_api_dotnet_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
