+++ mottu-mapping-api-dotnet.sln
@@ -1,24 +1,84 @@
+﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.5.2.0
+VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "mottu-mapping-api-dotnet", "mottu-mapping-api-dotnet.csproj", "{879CEFF3-0C2C-7951-3CFD-3C74BE3826D9}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Src", "Src", "{827E0CD3-B72D-47B6-A68D-7590B98EB39B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infra", "Src\Infra\Infra.csproj", "{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Src\Domain\Domain.csproj", "{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "Src\Application\Application.csproj", "{1F9B7C86-7934-4581-8F01-92B4A5B72F13}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApi", "Src\WebApi\WebApi.csproj", "{D2E5935E-D994-4D18-8F86-4C0539F31072}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{879CEFF3-0C2C-7951-3CFD-3C74BE3826D9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{879CEFF3-0C2C-7951-3CFD-3C74BE3826D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{879CEFF3-0C2C-7951-3CFD-3C74BE3826D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{879CEFF3-0C2C-7951-3CFD-3C74BE3826D9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Debug|x64.Build.0 = Debug|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Debug|x86.Build.0 = Debug|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Release|x64.ActiveCfg = Release|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Release|x64.Build.0 = Release|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Release|x86.ActiveCfg = Release|Any CPU
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3}.Release|x86.Build.0 = Release|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Debug|x64.Build.0 = Debug|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Debug|x86.Build.0 = Debug|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Release|x64.ActiveCfg = Release|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Release|x64.Build.0 = Release|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Release|x86.ActiveCfg = Release|Any CPU
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB}.Release|x86.Build.0 = Release|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Debug|x64.Build.0 = Debug|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Debug|x86.Build.0 = Debug|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Release|x64.ActiveCfg = Release|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Release|x64.Build.0 = Release|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Release|x86.ActiveCfg = Release|Any CPU
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13}.Release|x86.Build.0 = Release|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Debug|x64.Build.0 = Debug|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Debug|x86.Build.0 = Debug|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Release|x64.ActiveCfg = Release|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Release|x64.Build.0 = Release|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Release|x86.ActiveCfg = Release|Any CPU
+		{D2E5935E-D994-4D18-8F86-4C0539F31072}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8ADCDE56-FF04-4EEC-97D0-27E2C4668010}
+	GlobalSection(NestedProjects) = preSolution
+		{64A70200-4D72-4BC3-9D13-DF8393B9D6D3} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{1FEC8169-70F0-4EF1-9891-49A1155CE6EB} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{1F9B7C86-7934-4581-8F01-92B4A5B72F13} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{D2E5935E-D994-4D18-8F86-4C0539F31072} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 EndGlobal
