+++ src/Domain/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace Domain;
+
+public class Class1
+{
+
+}
