+++ src/Infra/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace Infra;
+
+public class Class1
+{
+
+}
