+++ src/WebApi/WebApi.http
@@ -1,0 +1,6 @@
+@WebApi_HostAddress = http://localhost:5272
+
+GET {{WebApi_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
