+++ src/WebApi/appsettings.Development.json
